requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
livekit==0.10.0
livekit-plugins-opencv==0.10.0
livekit-plugins-silero==0.10.0

# Web framework for UI
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
jinja2==3.1.2
aiofiles==23.2.1

# Telemetry and serial communication
pyserial==3.5
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
asyncio-mqtt==0.16.1

# Jetson specific (if needed)
# jetson-stats  # Uncomment if using jetson-stats for system monitoring
